requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.17.7
botocore==1.20.7
certifi==2020.6.20
cfe==0.0.15
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.11.0
django-storages==1.11.1
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
mailchimp-marketing==3.0.14
paywix==1.3.0
Pillow==7.1.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
urllib3==1.25.10
